Add a set of special cases to the XML_unparser, to deal with filename and line_number. I'm aware that this should really be fixed, rather than patched over, but this is done, and doing it properly isnt high on my priority list. Ideally, line_numbers should use Integer*s and filenames should never be missing.
A worry I had was that if files had weird line numbers and such, that a second reparse of the xml would lead to errors. Added this to the tests, but no such problem. The first AST should have line_numbers of zero, and the second should have Integer*s with value of zero, but this doesnt materialize (or .equals allows this??). Left the test in anyway.