BasicParseTest was in a file named basic_parse_tests. Renamed to basic_parse_test.php.
We no longer use the compile-time-include test, since its built in.

Include_and_co.php was written to test the parser, but it tries to include files that dont exist. This is tested in plenty of other places, so removed the test.

Since we moved errors to a different directory, include_massive shouldnt include some files.