Add a check for separation in write_var. If the array we are about to write into needs is in a copy-on-write set, we need to separate it.
This involves reordering functions, which makes the diff larger than it needs to be.

Fixed up the comment on separate_array to make it clear why we use separate_var in write_array, instead of separate_array.

If we look for an array to index, but get EG(uninitialized_zval_ptr), return that. We dont want to turn it into an array.