Minor change: If the function name passed is NULL, dont overwrite the EG( function_state_ptr), since it segfaults in some places.
Major change:
It has come to my attention that there is a reason for all these zval**s. After looking something up in a hashtable, the zval** points to the bucket of the hashtable. It is possible to overwrite this bucket, and indeed this happens in the case of passing by reference. this commit splits the fetch_var_arg and fetch_array_arg functions into _by_ref and normal versions. This actaully breaks more tests than it fixes, but its on the right track.

This means much of the separation is unnecessary. Instead of writing back into a hashtable, since we already have a pointer itno it, we just need to clone.

In the future, read_var and read_array should probably use this technique, and it may be possible to reduce the work of separate substantially.