I was trying to reduce factorial.php. The error only occured when using the debug_zval plugin. Factorial has flow-control in it, so it isnt possible to parse the reduced output. Additionally, removing some statements leads to infinite loops.
I had thought it would be OK to just run the output from misc/comp (it leaves a factorial.hir.php), but that has flow control in it, and we cant parse gotos and labels. It would be somewhat weird to try to reduce the uppered script.

The net result is that we now pass around XML, and unparse it when necessary. So to reduce and compile we use --read-ast-xml, and we unparse and upper it for PHP. We end by dumping the xml, lowered PHP and uppered PHP.

To get around the infinite loops, commands run are now non-blocking (stream_set_timeout apparently doesnt work too well), and they timeout after 4 seconds, which seemed a decent threshold.

Also, add -r and -d options, the same as misc/comp, to simplify reducing refcount and eval tests.