Abstract the reduce main step into a separate function. I think we can make the algorithm better, so its good to split the algorithm from the implementation.
In doing this, I noticed that when we remove a goto, for example, we'll start getting an error when we compile. The reducer will keep this error in, and reduce the script to something like:

<?php
	goto L4:
?>

which isnt useful.

So instead, we take a note of the labels of removed statements, and remove all statements which reference those labels. This appears to work on the test I was working on.