$a++ did too much work. It generated 4 statements, where only two were really necessary (and even then, not always).
I handled half of this in my previous commit, but there were still extra unnecessary statements being generated, and those did work (ie the LHSs weren't marked as unused).

This converts plain $a++ to ++$a, removes lone variables at an earlier stage (before the shredder can add a lhs to them), and does the pre- and post-op transformations earlier, so that they are already converted by the time the Desugar runs.