The test framework is now asynchronous. Compiled_vs_interpreter now launches programs asynchronously, then polls for them to finish. The advantage is that they can run two at once.
Since im not sure how to check for how many cores a machine has, ive limited it to only running one program at a time. However, the sweet spot on my machine is to have 3 running at once, so I suspect two would work fine for single core machines.

In addition to running one per core, this should have benefits comparable to make -j. By doing real work while another program is doing IO, we should see a speedup. In addition, there appears to be some time between when a program ends and PHP reports it as ended using proc_status. This should avoid waiting for that as well.

For now though, time is wasted polling, which should be avoided. This should get faster with more tuning.

Finally, this turns the compiled_vs_interpreted test into a simple data driven one. The other tests should be able to reuse this, and this should clear up the mess that a few tests are in right now.