Some revisions dont build. This adds a column in complete to indicate success or failure of the test.
It also refactors the tests so that failures are gracefully handled. Failures fall into 3 catagories:

script errors, which involve mailing the maintainer (currently me instead of phc-internals, during teething)
test errors, where the build wont run or the test breaks, and
svn errors, where the server is offline (we wait 5 minutes and try again).

The test should save its output on failure, and given a test error, should record the error and continue.