REmove the source_rep from the AST. Instead, we use a simpler variation, where the source_rep is an attribute, and so is automatically handled along with the other attributes. To do this, literals and casts inherit from a Source_rep node, which provided get_source_rep and set_source_rep.
This makes unparsers simpler, and the XML simpler aswell. However, it uncovers a bug in the AttrMap::get creates an attribute, with a value of NULL. This isnt allowed by assert_valid, but is explicitly allowed elsewhere. I'll settle for disallowing it everywhere, and fixing AttrMap (note that its still allowed in the XML_unparser, where its useful to have it be a little more resilient so that we can use it from debugging).