The replaces --read-ast-xml with --read-xml=pass_name, allowing us to parse XML and continue the compiler from any pass.
Having already removed source_rep, it is not hard to expand the XML_parser to
work on every IR in the same manner. By adding an {AST,HIR,MIR}_XML_parser, any
code duplication is avoided.

The parse function no longer uses a parameter to check if XML should be used.
Instead, an AST_XML_parser is allocated which parses as necessary.

Initializing and shutting down Xerces is now moved to main ().

A lot of code is moved from XML_parser.cpp to XML_parser.h. This is due to the
templatizing, and I don't believe it is avoidable.

A small amount of useful scaffolding is added to the pass_manager.

A test is added to check the error message if no test is found.