Rename loc() to pp_node(). Copy the dotty generation style from saturn, where we create dotty_nodes and dotty_edges on demand from the CFG.
Add a source_rep to the dotty nodes, so I can see what's going on. It seems to be hard to do generics in calypso. I've added an 'any' type, and source_rep() and attr() use it.

There is a small problem remaining. For example, for global_id{X}s, we need a 'source_rep(any{statement_Global{global_id{X}}}, SOURCE)', but we have source_rep(any{global_id{X}}, SOURCE), which means we need to add a predicate to infer the former from the latter. This isnt very generic at all, and is very annoying. I'll see if i can find a workaround.