Move the dot generation into cfgdot.clp.
Make sure that the extra source_reps are added into the MIR session, and not the current session.

Simplify do_cfg, by creating the CFG in the cfg session, instead of adding it to the cfg session afterwards.

Remove the argument from the mir, hir, and ast session declaration. This makes it possible to look them up without having to find the filename they were created in.