Start using Boost. Since using the Boost library in the dataflow branch, I've realized how crippled we are without it.
This adds a configure time check for boost, and tests that it works by replacing a few for_lci calls with calls to BOOST_FOREACH. The resulting code is simpler and shorter.

We also have immediate use for lexical_casts, and long term use for hash_maps and program options. The dataflow branch also uses Graph and Tribool.