Refactor the SSA building. Create Phi, Dominance, SSA_renaming and expand Basic_block.
Add Map, which is now the base for AttrMap (this should move to trunk). I has the convenient has(key) function.

Allow VARIABLE_NAME to be an SSA name. This seems preferable to creating a new SSA_NAME, and putting it into the MIR (or, heaven forbid, add a new SSA IR). A VARIABLE_NAME can only be turned into SSA form once, and is given a version.

Set now uses VARIABLE_NAME*s instead of strings. I added a comparison functions so that it wouldn't just compare based on pointer equality.

A lot of tidying up and reorganization.