Add a call to convert to SSA form, and fix some of the process. The Boost graph library is difficult to use, but this gets us most of the way there (using the internal index_map, rather than one specially made for us, as in Boost docs).
The next bug to fix is that some nodes do not have dominators. In a CFG, only the entry node should not have a dominator. However, unreachable nodes have no predecessor, and so they (and their successors) have no dominators. Cleaning up the CFG, removing unreachable nodes, will fix this.