Add if-simplification.
The total order on VARIABLE_NAME*s, used by use-def chains and Sets, was wrong. I'm not sure exactly what was wrong, but I've replaced it with a far simpler test (since versions are unique, just sort on versions in SSA form, and strings in non-SSA, and ban mixinging and matching SSA/non-SSA form in the same table).

Add dump() debugging functions to a number of classes.