Remove constant-folding and Remove_concat_null optimizations. They should be done by the optimizers.
Add an edge from Return nodes to the exit block.

when rebuilding SSA, rebuild dominance information.

Make a special case when replacing a list of nodes with a list including itself as the final node. This case applies for Branch_blocks and Exit_blocks.

When removing a basic block, remember to fix its phi nodes.

When removing unreachable nodes (tidy_up), use a mark-sweep style, makring nodes which are reachable from the entry node, and removing others. Before that, there was a problem with cycles.

Handle foreach_* nodes.

Fix PHP::is_true() - it returned true always.