Refactor all the CFG updating code. This is much less dense, much simpler, and much more likely to be correct. It renames functions to more accurately reflect their purpose, and refactors duplicated code out. There were a number of places where the duplicated code was out-of-sync, and those are probably the cause of the bugs.
Fix the SCCP. A change in lattice height of any defintion should add the SSA_edges of the definition for re-calculation. However, this was only is place for TOP->const, and only in Assign_var. This adds it for SSA_pre_op, visit_phi, and any assign_var. This keeps the iteration going until it actually converges. This fixes a big bug in bench_sieve, where a loop's back edges were destroyed, and lots of BBs were not post-dominated by the exit node.