Currently, the largest file in the phc testsuite is 2000 lines, and this requires 150MB of RAM. This is unacceptable, since we should be able to compile an entire web-app, of, say, 50000 lines, without hitting a problem. Obviously, we need to fix the GC problems, and enable it by default.
Having studied the libgc documentation, the most likely cause of the segfaults we experience is that memory is still pointed to by objects not traced by the collector. I've added a plan for fixing this. One of the biggest problems was a simple one: Object required clone() to be implemented. So far, we've only needed deep copies on AST/HIR/MIR objects and Lists. So we need a root object which doesnt require clone(). This is now GC_obj. Object still requires clone(), and it will still be the root for maketea generated sources. All other classes we define will be migrated to GC_obj.