Handle assign_var_expr_var by generating a list of CODE. So, not a very large step. I was pretty much forced to add method to the definition, although that's worked out alright.
Its very easy to drop a bit of C code, but fortunately something always fails in that case.

Comments and scope braces have moved a little bit more locally, which is a small bit more elegant. However, they had to be duplicated between for UNINTERPRETED, which is certainly not more elegant.