Change the lowering for post-increment operators. Rather than changing
while($i++ < 10)

to

<?php
        while (True) {
                $TS0 = $i;
                ++$i;
                if (!($TS0 < 10)) {
                        break;
                }
                print($i);
        }
?>

we change it to

<?php
        while (True) {
                if (!($i < 10)) {
                        ++$i;
                        break;
                } else {
                        ++$i;
                }
                print($i);
        }
?>

This avoids the separation of $i in the loop body, which was a significant
performance drain in the strcat benchmark. 

Unfortunately, this introduces one new bug into phc; we don't currently handle

  return $i++;

correctly. It gets turned into

  return $i;
  $i++;

which is obviously incorrect. This shows in the isset_eval_order test and also
in a new post_increment_in_return test which tests this explicitly. This should
be handled in a pass before the Pre_post_op_shredder, which should add
temporaries to return statements with side-effecting arguments.

Also discovered a latent bug in phc and added a test for it
(post_op_with_side_effect). Statements of the form $i[f()]++ cause f() to be
called twice in phc. This can be fixed by adding temps for the call to f() here
too (the pass mentioned in the previous paragraph can be used here too).