Since Def_use is outside of MIR::Statements, there isnt really a good reason to use MIR::VARIABLE_NAMEs for SSA. Alias_names seem like much better options.
Unfortunately, using VARIABLE_NAMEs is very much ingrained. This replaces all uses of VARIABLE_NAMEs with Alias_names, for SSA and use-def information.

There were a few spots where this was tricky, so I've commented them out, or add a phc_TODO(). In a few places, it is clearly wrong, as copies are used for Alias_names, but I'll have to deal with that later.

This is the first occasion in which it's compiled and linked, so its a good time to commit. It doesnt work yet though. Hopefully I wont have to revert this.