The old SSA representation, based on the VARIABLE_NAMEs in the statements themselves, was vey complicated. IT doesnt need to be so complicated, and using the Def_use analysis run during WPA, its complexity should be reduced.
This first part copies the Alias_names to Def_use_web, where defs and uses will be stored outside of the MIR nodes.

Meanwhile, the old SSA representation has had all its components prefixed with 'old_'. Its far too big a job to remove them all now, and I need to make sure that all of the odd edge cases are taken care of. The aims over the next few commits is to rebuild a much simpler SSA form, and to remove (or integrate) anything with old_ in its name.