Remove methods which are not called.
It should be clear here that the optimization framework, as-is, does not optimize libraries/extensions very well (by which I mean that it only really works for programs with a __MAIN__ function).

I imagine it wont take too much work to allow that (at least for simple cases), but its not a priority for now.