Implement iteration for the whole-program analysis. This iterates until its whole program solution is the same as the last time. This means implementing equals() for each analysis.
To simplifiy that, I implemented .equals for Sets and Maps (couldnt use == because that doesnt deal check pointers (hmmm, there might be a way around that...)). This means copying the solution I used for cloning. The only addition is support for containers of non-pointers, which will call operator==() if .equals is not defined (or more specifically, if SET_COMPARATOR is not set).