Refactor Points_to, enforcing further the separatino between references, points-to edges and fields. The API is now totally separated, and generic functions are mostly removed. They were the cause of a lot of confusion, and subtle bugs, so things should be a lot clearer now. A few things which didn't make any sense have already been spotted and removed.
Removed unused functions.
Re-implement edge manipulation functions with type-safety.
Add a lot more type safety, and some const-correctness (I would like to make edges and nodes in the points-to graph consts, but that permiates to a lot of places, so I'll leave that until later).

get_references () now returns a pair which includes the certainty. This will allow us the fix the awful logic for determining if a definition is killing, and allow references to work properly.