There were a few places where we created an array, then set a type for its values. Refactor that out into assign_path_typed_array.
Start fixing casts so that we avoid duplicating anything with coerce_to_string. This currently breaks anything with a cast, but it will be fixed soon.

Finally, this adds a flag to all the assign_path_* assignments, as to whether or not to kill. This allows a simple merging of results (for example, some functions have two cases, and its annoying to try and replicate the assign_path_* stuff on Index_nodes).