Fix a number of problems involving stacks.
I was using a local variables for a number of tasks:
	- the unique_count for FAKE variables
	- passing information from an assign_var to its expression
	- passing block_cx to the CFG visitor

I should have been using a stack for each of these, so that the result wouldn't be overwritten depending on what happens in a caller. This also solves the problem of the numbering of fake nodes.

This also adds a few 'destroy_fake_nodes' calls at points where fake nodes shouldn't exist.

Some miscellaneous changes:
	- FAKE is made abstract so that it will always appear, and won't vanish without my knowledge.
	- When remove_storage_node() is called, it should always erase storage nodes, not just when called from dec_inc ().
	- Print symbol tables and storage nodes (which have no edges from them) in the points-to graph.