Instead of converting
	$x = array (5, 6);

into (simplified)

	$x[] = 5;
	$x[] = 6;

convert it into 

	$x[0] = 5;
	$x[1] = 6;


This makes the analysis much simpler and more precise, without having to track array_nexts, which would be trickey (especially when copying).

It would be worthwhile to convert them to

	$x = array ();

instead of the unsetting and copying we do, but I havent got time (though I should really have done it long ago).