Up until now, we'd been creating the Def Use Web from the Def Use and using the DUW to convert the defs and uses to SSA form with HSSA, but we weren't updating the DUW and parts of it were still in non-SSA form after convert_to_hssa_form was run.
The solution to this is to build the web again with the new variable names, which are stored in the DUW's defs and uses at the end of convert_to_hssa_form ().

I've modified build_web () so that it now takes a boolean 'update' argument, which, if true, clears named_defs/uses and def/use_ops and builds them again in SSA form, whereas if it's false, it obtains defs and uses from the Def Use.