Do not use unique names for each new array in a block. This makes the results less precise, but the other way we got infinite loops, the analysis might not have converged, and we ended up with hundreds of arrays being modelled in some cases.
Since we no longer tack a number on to the end of an object's name, that object would get the name of the constructor. Stick SYM onto the front to prevent that.

It isnt correct to just start using an abstract array. We need to say that all of its fields are potentially null.